Delete the display method completely
The display was not working correctly because once the event listeners
were successfully created, when clicked, the buttons were calling
the method to simply read user input instead of the method that reads
and displays it.

A potential solution might be to incorporate both clicking and
displaying into the same function.